change the way RemainingElements calculates and remove safety from MoveNext/Previous
This is mainly to fix a problem where it's not exactly intuitive to code
a loop which runs through a buffer using MoveNext() up until the end of
RemainingElements. When the loop ends, you'd expect to be able to take
that same buffer, Extend() it, and be able to immediately start running
through it again without calling MoveNext() one extra time inbetween
both loops. This fix makes it so this kind of "intuitive" usage pattern
works, at the expense of no automatic bounds safety with CurrentPosition